Stopwatch-With-JS/index.html
Learned a few new variables and how to use them.
Learned how to pass a variable as 00.
Learned how to calculate number of minutes/number of seconds.
Learned about ternary operators.
Learned the setInterval function.
Learned about data attributes and that we can attach them to our html.
Successful in learning to create a stopwatch!